pyproject.toml
[project]
name = "hybrid-log-classification-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair<5",
    "fastapi>=0.128.4",
    "groq>=1.0.0",
    "joblib>=1.5.3",
    "matplotlib",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.2",
    "standard-imghdr>=3.13.0",
    "streamlit>=1.19.0",
    "uvicorn>=0.40.0",
    "xlsxwriter>=3.2.9",
]
